requirements.txt
streamlit
python-dotenv
azure-cognitiveservices-vision-computervision
google-cloud-vision
google-generativeai 
ollama